% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_distance_SW.R
\name{classify_distance_SW}
\alias{classify_distance_SW}
\title{Distance classifier with the sliding window method}
\usage{
classify_distance_SW(x, label_movements)
}
\arguments{
\item{label_movements}{the list of possible labels for movements}

\item{x}{the table of treated data, according to the criteria which will be used to classify}

\item{criteria}{the criteria allowing to classify the data}
}
\value{
Return a list of labels of the sequence
}
\description{
This function classifies the data segment and finds the labels thanks to the distance.
}
\examples{
classify_distance_SW(x, labels)
}
