% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_distance_without_SW.R
\name{classify_distance_without_SW}
\alias{classify_distance_without_SW}
\title{Distance classifier without sliding window}
\usage{
classify_distance_without_SW(descriptors_distances, label_movements)
}
\arguments{
\item{descriptors_distances}{the vector of distances}

\item{label_movements}{the lit of movement labels}
}
\value{
Return the label corresponding to this movement
}
\description{
This function classifies the data segment and finds the label thanks to the distances.
}
\examples{
classify_distance_without_SW(descr_dist, labels)
}
