% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_forest_without_SW.R
\name{classify_forest_without_SW}
\alias{classify_forest_without_SW}
\title{Forest Classifier for a method without sliding window}
\usage{
classify_forest(descriptors_database, descriptors, no_labelled_threshold, tree_number)
}
\arguments{
\item{descriptors_database}{the table of database descriptors}

\item{descriptors}{the vector of descriptors to classify}

\item{no_labelled_threshold}{the threshold to determine if the movement label is known or not}

\item{tree_number}{the number of tree in the random forest}
}
\value{
Return a list of labels of the sequence
}
\description{
This function classifies the data segment and finds the labels thanks to a random forest
}
\examples{
classify_forest(database, descriptors, threshold, n_tree)
}
