% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_descriptors_db.R
\name{compute_descriptors_db}
\alias{compute_descriptors_db}
\title{Database descriptors computation}
\usage{
compute_descriptors_db(x, criteria)
}
\arguments{
\item{database}{the database}

\item{criteria}{the method of classification}
}
\value{
Return a table with the descriptors for each label
}
\description{
This function computes the descriptors for each label on the database.
}
\examples{
compute_descriptors_db(x, distance)
}
