% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_descriptors_window.R
\name{compute_descriptors_window}
\alias{compute_descriptors_window}
\title{Window descriptors computation}
\usage{
compute_descriptors_window(left_window_data, right_window_data, width_sliding_window, descriptors_database, criteria)
}
\arguments{
\item{left_window_data}{the window of left wrist data}

\item{right_window_data}{the window of right wrist data}

\item{width_sliding_window}{the size of the window in number of data}

\item{descriptors_database}{the table of descriptors of the database}

\item{criteria}{the way of classification}
}
\value{
Return the vector of descriptors of the window
}
\description{
This function computes the descriptors of the window which slides on the dataset to treat.
}
\examples{
compute_descriptors_window(left_data, right_data, size_wind, descr_database, "distance")
}
