% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_distances.R
\name{compute_distances}
\alias{compute_distances}
\title{Distance between descriptors vectors of the window and the database computation}
\usage{
compute_distances(descr, database_descr, label, distance_threshold)
}
\arguments{
\item{normalized_descriptors}{the window descriptors vector}

\item{normalized_descriptors_database}{a label descriptors vector}

\item{label_movements}{the liste of movements labels}

\item{distance_threshold}{the no motion distance threshold}
}
\value{
Return a vector of distances between the window descriptors vector and each label
}
\description{
This function computes the distances between the window descriptors vector and all the database descriptors vector, in order to obtain for the window, a distance from each label.
}
\examples{
compute_distances(x, y, label, dist_thres)
}
