% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_occurrences.R
\name{count_occurrences}
\alias{count_occurrences}
\title{Label occurrences computation}
\usage{
count_occurrences(label_movements, labels)
}
\arguments{
\item{label_movements}{the list of labels of the database}

\item{labels}{the table of labels of the window}
}
\value{
Return a table with the occurrences for each label
}
\description{
This function counts the occurrences of each movement label of the labels list and creates a table with a column of labels and a column with the occurrences.
}
\examples{
count_occurrences(labels, segment_list)
}
