% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean_distance.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Euclidean distance between vectors computation}
\usage{
euclidean_distance(x, y)
}
\arguments{
\item{x}{the first vector}

\item{y}{the second vector}
}
\value{
Return the euclidean distance between the two vectors
}
\description{
This is a function which computes the euclidean distance between the two vectors entered as arguments.
}
\examples{
euclidean_distance(c(12, 6, 7), c(45, 87, 19))
euclidean_distance(c(3, 4), c(1, 2))
}
