% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_MAE.R
\name{evaluate_MAE}
\alias{evaluate_MAE}
\title{Algorithm performance evaluation}
\usage{
evaluate_MAE(x, y, db)
}
\arguments{
\item{predicted_labels}{the obtained result, a table with the movement labels and the occurrence of each label}

\item{expected_seq}{the expected result, a table with the movement labels and the occurrence of each label}

\item{database}{the database}
}
\value{
Return the MAE between the expected result and what it's obtained.
}
\description{
This function computes the MAE, which is the mean of the absolute error between the expected result and the obtained result.
}
\examples{
evaluate_MAE(x, y, database)
}
