% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immobility_withdrawal_simplified.R
\name{immobility_withdrawal_simplified}
\alias{immobility_withdrawal_simplified}
\title{Simplified immobility withdrawal}
\usage{
immobility_withdrawal_simplified(resampled_accelerations, resampling_frequency, immobility_threshold, time_threshold)
}
\arguments{
\item{resampled_accelerations}{the table of resampled accelerations data}

\item{resampling_frequency}{the resampling frequency}

\item{immobility_threshold}{the immobility threshold}

\item{time_threshold}{the first filter time threshold}

\item{min_motion_length}{The minimal length of motion}
}
\value{
Return a table with no more immobility and a column which indicates the first segmentation of the dataset
}
\description{
This function withdraws all the data which are smaller than the immobility threshold, allowing to keep only the data corresponding to movement. It keeps the data only if the segment is large enough, thanks to a first filter segment size threshold. This is a first filter which allows to segment a first time the dataset.
}
\examples{
immobility_withdrawal_simplified(x, 2, 7, 1)
}
