% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{merger}
\alias{merger}
\title{Merger of labels for the sliding window approach}
\usage{
merger(predicted_labels)
}
\arguments{
\item{predicted_labels}{the table of predicted labels}
}
\value{
Return a table of the labels merged. Their is a single label for a motion segment.
}
\description{
This function merged the labels thanks to criteria.
}
\examples{
merger(labels)
}
