% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Data resampling}
\usage{
resample(left_wrist, right_Wrist, resampled_frequency)
}
\arguments{
\item{left_wrist}{the first dataset containing timed data of the left wrist}

\item{resampling_frequency}{the resampling frequency used to resample the timed data}

\item{right_Wrist}{the second dataset containing timed data of the right wris}
}
\value{
Return a table with the resampled overall accelerations from the two wrist
}
\description{
This function resamples the timed data from two dataset at the resampling frequency entered as input.
}
\examples{
resample(tab1, tab2, 100)
}
