% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show}
\alias{show}
\title{Display of graphics}
\usage{
show(data, real_data, step)
}
\arguments{
\item{data}{The table of motions found with machine learning}

\item{real_data}{The table of real motions}

\item{step}{The type of graph wanted according to the step of the treatment}
}
\value{
Return a graph
}
\description{
This function display the absolute sum of accelerations for each movement : the rhombus represent the motions found with the machine learning and the point represent the real motions.
}
\examples{
show(data, real_data, step)
}
